sm1000_main: Add new menu system.
This is a first attempt at a proper UI for the SM1000.  The SELECT and
BACK buttons now have two functions.

When the device powers on, the SELECT/BACK buttons cycle between the
operating modes when pressed momentarily.

Hold down SELECT and the POWER LED starts flashing, now you are in MENU
mode.  SELECT/BACK now move between the menu items, holding down SELECT
chooses an item, and holding down BACK backs out without making a
change.

The menu structure at the moment is as follows:

-> Root menu:
	|---> "MODE": Operating mode
	|	|---> "ANA": 	Analogue mode
	|	|---> "1600":	FreeDV 1600
	|	'---> "TONE":	Test tone
	|
	'---> "UI": User interface options
		|---> "FREQ":	Morse frequency
		|		50-2000 Hz in 50Hz steps
		|---> "WPM":	Morse send speed (PARIS standard)
		|		5-60 WPM in 5 WPM steps
		'---> "VOL":	Menu volume
				1-15 units.

git-svn-id: https://svn.code.sf.net/p/freetel/code/codec2-dev@2370 01035d8c-6547-0410-b346-abe4f91aad63